Move all processors under one interface
