--- app/src/main/java/com/babylon/wallet/android/presentation/transaction/analysis/GeneralTransferAnalysis.kt
@@ -1,119 +1,0 @@
-package com.babylon.wallet.android.presentation.transaction.analysis
-
-import com.babylon.wallet.android.domain.model.resources.Resource
-import com.babylon.wallet.android.domain.usecases.ResolveDAppInTransactionUseCase
-import com.babylon.wallet.android.domain.usecases.transaction.GetTransactionBadgesUseCase
-import com.babylon.wallet.android.presentation.transaction.AccountWithTransferableResources
-import com.babylon.wallet.android.presentation.transaction.PreviewType
-import com.radixdlt.ret.ExecutionSummary
-import com.radixdlt.ret.ResourceIndicator
-import rdx.works.profile.data.model.pernetwork.Network
-import rdx.works.profile.domain.GetProfileUseCase
-import rdx.works.profile.domain.accountsOnCurrentNetwork
-import rdx.works.profile.domain.defaultDepositGuarantee
-
-// Generic transaction resolver
-suspend fun ExecutionSummary.resolveGeneralTransaction(
-    resources: List<Resource>,
-    getTransactionBadgesUseCase: GetTransactionBadgesUseCase,
-    getProfileUseCase: GetProfileUseCase,
-    resolveDAppInTransactionUseCase: ResolveDAppInTransactionUseCase
-): PreviewType {
-    val badges = getTransactionBadgesUseCase(accountProofs = presentedProofs)
-    val dApps = resolveDApps(resolveDAppInTransactionUseCase).distinctBy {
-        it.first.definitionAddresses
-    }
-    val involvedAccountAddresses = accountWithdraws.keys + accountDeposits.keys
-    val allAccounts = getProfileUseCase.accountsOnCurrentNetwork().filter {
-        it.address in involvedAccountAddresses
-    }
-
-    val defaultDepositGuarantee = getProfileUseCase.defaultDepositGuarantee()
-
-    return PreviewType.Transfer.GeneralTransfer(
-        from = resolveFromAccounts(resources, allAccounts),
-        to = resolveToAccounts(resources, allAccounts, defaultDepositGuarantee),
-        badges = badges,
-        dApps = dApps
-    )
-}
-
-private fun ExecutionSummary.resolveFromAccounts(
-    allResources: List<Resource>,
-    allAccounts: List<Network.Account>
-) = accountWithdraws.map { withdrawEntry ->
-    val transferables = withdrawEntry.value.map {
-        it.toWithdrawingTransferableResource(
-            resources = allResources,
-            newlyCreatedMetadata = newEntities.metadata,
-            newlyCreatedEntities = newEntities.resourceAddresses
-        )
-    }
-
-    val ownedAccount = allAccounts.find { it.address == withdrawEntry.key }
-    if (ownedAccount != null) {
-        AccountWithTransferableResources.Owned(
-            account = ownedAccount,
-            resources = transferables
-        )
-    } else {
-        AccountWithTransferableResources.Other(
-            address = withdrawEntry.key,
-            resources = transferables
-        )
-    }
-}
-
-/**
- * The account deposits order that comes from RET does not take into account the order we maintain in the app.
- * This method sorts accounts in the same order they are appearing in the dashboard.
- * TODO revisit if this can be done using Comparator.comparing { item -> allAccounts.map { it.address }.indexOf(item) }
- */
-fun Map<String, List<ResourceIndicator>>.sort(allAccounts: List<Network.Account>): Map<String, List<ResourceIndicator>> {
-    val allAccountsAddresses = allAccounts.map { it.address }
-
-    // Only account deposits that we own
-    val ownedAccountDeposits = this.toList().filter {
-        allAccountsAddresses.indexOf(it.first) != -1
-    }
-
-    // Sorted owned accounts deposits according to the all accounts order
-    val ownedAccountDepositsSorted = ownedAccountDeposits.sortedBy {
-        allAccountsAddresses.indexOf(it.first)
-    }.toMap()
-
-    // The rest of account deposits (ones we do not own)
-    val thirdPartyAccountDeposits = this.minus(ownedAccountDepositsSorted.keys)
-
-    return ownedAccountDepositsSorted.plus(thirdPartyAccountDeposits)
-}
-
-private fun ExecutionSummary.resolveToAccounts(
-    allResources: List<Resource>,
-    allAccounts: List<Network.Account>,
-    defaultDepositGuarantees: Double
-): List<AccountWithTransferableResources> {
-    return accountDeposits.sort(allAccounts).map { depositEntry ->
-        val transferables = depositEntry.value.map {
-            it.toDepositingTransferableResource(
-                resources = allResources,
-                newlyCreatedMetadata = newEntities.metadata,
-                newlyCreatedEntities = newEntities.resourceAddresses,
-                defaultDepositGuarantees = defaultDepositGuarantees
-            )
-        }
-
-        val ownedAccount = allAccounts.find { it.address == depositEntry.key }
-        if (ownedAccount != null) {
-            AccountWithTransferableResources.Owned(
-                account = ownedAccount,
-                resources = transferables
-            )
-        } else {
-            AccountWithTransferableResources.Other(
-                address = depositEntry.key,
-                resources = transferables
-            )
-        }
-    }
-}
